Adding Custom Exceptions

+++ src/Core/HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType
@@ -23,6 +24,14 @@ namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeave
             // retrieve  domain object
             var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
+            //verify if the object exists
+            if (leaveTypeToDelete == null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
+
+
+            //remove from DB
             await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
 
             // return record id
